config_editor 코드 리팩토링 및 디자인 변경
