+++ kiosk-builder-app/run_gui.py
@@ -34,10 +34,32 @@ else:
 os.chdir(program_directory)
 
 def show_settings_window():
-    from config_editor import ConfigEditor
-    settings_window = ConfigEditor()
-    settings_window.show()
-
+    try:
+        print("설정 창을 열려고 시도 중...")
+        
+        from config_editor import ConfigEditor
+        settings_window = ConfigEditor()
+        settings_window.show()
+        
+        # 윈도우 객체가 가비지 컬렉션되지 않도록 전역 변수로 저장
+        global main_window
+        main_window = settings_window
+        
+        print("설정 창이 열렸습니다.")
+    except Exception as e:
+        print(f"오류 발생: {e}")
+        import traceback
+        traceback.print_exc()
+        
+        # 오류 메시지 표시
+        from PySide6.QtWidgets import QMessageBox
+        error_box = QMessageBox()
+        error_box.setWindowTitle("오류")
+        error_box.setText(f"설정 창을 여는 중 오류가 발생했습니다: {e}")
+        error_box.setDetailedText(traceback.format_exc())
+        error_box.setIcon(QMessageBox.Critical)
+        error_box.exec_()
+        
 if __name__ == "__main__":
     app_id = "kiosk_builder_app_unique_id"
     app = SingleApplication(app_id, sys.argv)
